# doc/api.yaml
---
openapi: 3.0.0
paths:
  "/":
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - App
  "/user":
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateUserDto"
      responses:
        '201':
          description: The user has been created.
        '400':
          description: Bad request. Body does not contain required fields.
      summary: Create user
      tags:
      - Users
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Successful operation
      summary: Get all users
      tags:
      - Users
  "/user/{id}":
    get:
      operationId: UserController_findOne
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      summary: Get single user by id
      tags:
      - Users
    put:
      operationId: UserController_update
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdatePasswordDto"
      responses:
        '200':
          description: The user has been updated.
        '400':
          description: Bad request. userId is invalid (not uuid)
        '403':
          description: oldPassword is wrong
        '404':
          description: User not found
      summary: Update a user's password
      tags:
      - Users
    delete:
      operationId: UserController_remove
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      summary: Delete user
      tags:
      - Users
  "/artist":
    post:
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateArtistDto"
      responses:
        '201':
          description: Artist created successfully
        '400':
          description: Bad request. Invalid data
      summary: Create artist
      tags:
      - Artists
    get:
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: Return all artists
      summary: Get all artists
      tags:
      - Artists
  "/artist/{id}":
    get:
      operationId: ArtistController_findOne
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Return artist by id
        '400':
          description: Bad request. Invalid UUID
        '404':
          description: Artist not found
      summary: Get artist by id
      tags:
      - Artists
    put:
      operationId: ArtistController_update
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateArtistDto"
      responses:
        '200':
          description: Artist updated successfully
        '400':
          description: Bad request. Invalid data
        '404':
          description: Artist not found
      summary: Update artist information
      tags:
      - Artists
    delete:
      operationId: ArtistController_remove
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Artist deleted successfully
        '400':
          description: Bad request. Invalid UUID
        '404':
          description: Artist not found
      summary: Delete artist
      tags:
      - Artists
  "/album":
    post:
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateAlbumDto"
      responses:
        '201':
          description: Album created successfully
        '400':
          description: Bad request. Invalid data
      summary: Add new album
      tags:
      - Albums
    get:
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: Return all albums
      summary: Get all albums
      tags:
      - Albums
  "/album/{id}":
    get:
      operationId: AlbumController_findOne
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Return album by id
        '400':
          description: Bad request. Invalid UUID
        '404':
          description: Album not found
      summary: Get album by id
      tags:
      - Albums
    put:
      operationId: AlbumController_update
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateAlbumDto"
      responses:
        '200':
          description: Album updated successfully
        '400':
          description: Bad request. Invalid data
        '404':
          description: Album not found
      summary: Update album information
      tags:
      - Albums
    delete:
      operationId: AlbumController_remove
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Album deleted successfully
        '400':
          description: Bad request. Invalid UUID
        '404':
          description: Album not found
      summary: Delete album
      tags:
      - Albums
  "/track":
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTrackDto"
      responses:
        '201':
          description: Track created
        '400':
          description: Bad request
      summary: Add new track
      tags:
      - Tracks
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Success
      summary: Get all tracks
      tags:
      - Tracks
  "/track/{id}":
    get:
      operationId: TrackController_findOne
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Success
        '400':
          description: Invalid UUID
        '404':
          description: Track not found
      summary: Get track by id
      tags:
      - Tracks
    put:
      operationId: TrackController_update
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTrackDto"
      responses:
        '200':
          description: Track updated
        '400':
          description: Bad request
        '404':
          description: Track not found
      summary: Update track information
      tags:
      - Tracks
    delete:
      operationId: TrackController_remove
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Track deleted
        '400':
          description: Invalid UUID
        '404':
          description: Track not found
      summary: Delete track
      tags:
      - Tracks
  "/favs":
    get:
      operationId: FavsController_findAll
      parameters: []
      responses:
        '200':
          description: Success
      summary: Get all favorites
      tags:
      - Favorites
  "/favs/artist/{id}":
    post:
      operationId: FavsController_addArtist
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: Added successfully
        '400':
          description: Invalid UUID
        '422':
          description: Artist doesn't exist
      summary: Add artist to favorites
      tags:
      - Favorites
    delete:
      operationId: FavsController_deleteArtist
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid UUID
        '404':
          description: Artist not in favorites
      summary: Delete artist from favorites
      tags:
      - Favorites
  "/favs/album/{id}":
    post:
      operationId: FavsController_addAlbum
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: Added successfully
        '400':
          description: Invalid UUID
        '422':
          description: Album doesn't exist
      summary: Add album to favorites
      tags:
      - Favorites
    delete:
      operationId: FavsController_deleteAlbum
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid UUID
        '404':
          description: Album not in favorites
      summary: Delete album from favorites
      tags:
      - Favorites
  "/favs/track/{id}":
    post:
      operationId: FavsController_addTrack
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: Added successfully
        '400':
          description: Invalid UUID
        '422':
          description: Track doesn't exist
      summary: Add track to favorites
      tags:
      - Favorites
    delete:
      operationId: FavsController_deleteTrack
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid UUID
        '404':
          description: Track not in favorites
      summary: Delete track from favorites
      tags:
      - Favorites
info:
  title: RSS Home Library Service
  description: Home music library service - olenaweb/nodejs2025Q4-service
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: User login
          example: TestUser
        password:
          type: string
          description: User password
          example: Password123
      required:
      - login
      - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: Previous password
          example: oldPassword123
        newPassword:
          type: string
          description: New password
          example: newPassword456
      required:
      - oldPassword
      - newPassword
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: Artist name
          example: The Beatles
        grammy:
          type: boolean
          description: Does the artist have a Grammy award
          example: true
      required:
      - name
      - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: Artist name
          example: The Beatles
        grammy:
          type: boolean
          description: Does the artist have a Grammy award
          example: true
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: Album name
          example: Innuendo
        year:
          type: number
          description: Release year
          example: 1991
        artistId:
          type: string
          description: Artist ID
          example: 8d3ee94f-876d-4ba1-a64c-f8d6ace64302
          nullable: true
      required:
      - name
      - year
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: Album name
          example: Innuendo
        year:
          type: number
          description: Release year
          example: 1991
        artistId:
          type: string
          description: Artist ID
          example: 8d3ee94f-876d-4ba1-a64c-f8d6ace64302
          nullable: true
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: string
          description: Artist ID
          example: 8d3ee94f-876d-4ba1-a64c-f8d6ace64302
          nullable: true
        albumId:
          type: string
          description: Album ID
          example: '03019106-913d-41b8-8c88-b34b593f7d07'
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 355
      required:
      - name
      - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          description: Track name
          example: Bohemian Rhapsody
        artistId:
          type: string
          description: Artist ID
          example: 8d3ee94f-876d-4ba1-a64c-f8d6ace64302
          nullable: true
        albumId:
          type: string
          description: Album ID
          example: '03019106-913d-41b8-8c88-b34b593f7d07'
          nullable: true
        duration:
          type: number
          description: Track duration in seconds
          example: 355